refactor(fe): allow delete function to have extra arg
